updated code

+++ resources/views/articles/list.blade.php
@@ -4,7 +4,9 @@
             <h2 class="font-semibold text-xl text-gray-800 leading-tight">
                 {{ __('Articles') }}
             </h2>
-            <a href="{{ route('articles.create') }}" class="bg-slate-700 text-sm text-white rounded-md px-3 py-2">Create</a>
+            @can('create articles')
+                <a href="{{ route('articles.create') }}" class="bg-slate-700 text-sm text-white rounded-md px-3 py-2">Create</a>
+            @endcan
         </div>
     </x-slot>
 
@@ -31,8 +33,12 @@
                             <td class="px-6 py-3 text-left">{{ $article->author }}</td>
                             <td class="px-6 py-3 text-left">{{ \Carbon\Carbon::parse($article->created_at)->format('d M, Y')  }}</td>
                             <td class="px-6 py-3 text-center">
-                                <a href="{{ route('articles.edit',$article->id) }}" class="bg-slate-700 text-sm text-white rounded-md px-3 py-2 hover:bg-slate-600">Edit</a>
-                                <a href="javascript:void(0)" onclick="deleteArticle({{ $article->id }})" class="bg-red-600 text-sm text-white rounded-md px-3 py-2 hover:bg-red-500">Delete</a>
+                                @can('edit articles')
+                                    <a href="{{ route('articles.edit',$article->id) }}" class="bg-slate-700 text-sm text-white rounded-md px-3 py-2 hover:bg-slate-600">Edit</a>
+                                @endcan
+                                @can('delete articles')
+                                    <a href="javascript:void(0)" onclick="deleteArticle({{ $article->id }})" class="bg-red-600 text-sm text-white rounded-md px-3 py-2 hover:bg-red-500">Delete</a>
+                                @endcan
                             </td>
                         </tr>                        
                     @endforeach
